+++ packages/active-element/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @solid-primitives/active-element
 
+## 2.0.7
+
+### Patch Changes
+
+- Updated dependencies [c2866ea6]
+- Updated dependencies [c2866ea6]
+  - @solid-primitives/event-listener@2.2.6
+  - @solid-primitives/utils@5.0.0
+
 ## 2.0.6
 
 ### Patch Changes