+++ packages/raf/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @solid-primitives/raf
 
+## 2.1.6
+
+### Patch Changes
+
+- Updated dependencies [c2866ea6]
+  - @solid-primitives/utils@5.0.0
+
 ## 2.1.5
 
 ### Patch Changes