+++ packages/set/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @solid-primitives/set
 
+## 0.3.2
+
+### Patch Changes
+
+- @solid-primitives/trigger@0.0.4
+
 ## 0.3.1
 
 ### Patch Changes